pyproject.toml
[tool.poetry]
packages = [
    {include = 'src'},
]
name = "some-game"
version = "0.1.0"
description = ""
authors = ["A-V-tor <avtorca4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.112.2"}
sqlalchemy = "^2.0.33"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
sqladmin = "^0.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
